requirements.txt
django~=3.2.4
djangorestframework~=3.12.4
djangorestframework-simplejwt
drf-yasg
pillow
django-cors-headers
celery
redis
python-decouple
psycopg2-binary
django-ckeditor
django-filter
beautifulsoup4
lxml